moving Hitran / Hitemp / CDSD common parts to API
