+++ radis/io/__init__.py
@@ -5,17 +5,13 @@
 
 """
 
-
-from .cdsd import cdsd2df
 from .exomol import fetch_exomol
 from .geisa import fetch_geisa
 from .hitemp import fetch_hitemp
-from .hitran import fetch_hitran, hit2df
+from .hitran import fetch_hitran
 from .query import fetch_astroquery
 
 __all__ = [
-    "cdsd2df",
-    "hit2df",
     "fetch_hitemp",
     "fetch_hitran",
     "fetch_exomol",
