+++ radis/lbl/loader.py
@@ -59,23 +59,19 @@ import numpy as np
 import pandas as pd
 
 from radis import config
+from radis.api.cdsdapi import cdsd2df
 
 # from radis.io.geisa import gei2df
 from radis.api.hdf5 import hdf2df
+from radis.api.hitranapi import hit2df, parse_global_quanta, parse_local_quanta
 from radis.db.classes import get_molecule
 from radis.db.molecules import getMolecule
 from radis.db.molparam import MOLPARAMS_EXTRA_PATH, MolParams
 from radis.db.references import doi
-from radis.io.cdsd import cdsd2df
 from radis.io.exomol import fetch_exomol
 from radis.io.geisa import fetch_geisa
 from radis.io.hitemp import fetch_hitemp
-from radis.io.hitran import (
-    fetch_hitran,
-    hit2df,
-    parse_global_quanta,
-    parse_local_quanta,
-)
+from radis.io.hitran import fetch_hitran
 from radis.io.query import fetch_astroquery
 from radis.io.tools import drop_object_format_columns, replace_PQR_with_m101
 from radis.levels.partfunc import (
